Provide means to validate info provided by user during registration
This class delegates the task of validating info provided
by the user during registration to the individual validators,
passed in as collaborators (dependencies). One important
thing to note here is that the collaborators can be either a
composite or individual validator object. This gives us the flexibility
of composing multiple validators without modifying this class.