Wire up dependencies required by registration validator
ValidatorsFactory class is responsible for creating validators
and wiring up necessary dependencies. The client code for
these validators don't need to know which collaborators are
required by each validator to do their job. As a result, if we
add or remove collaborators in the future, the client code
remains unaffected.