Provide means to validate info entered by user in settings view
This class delegates the task of validating settings
info to the individual validators, passed in as collaborators
(dependencies). One important thing to note here is that
the collaborators can be either a composite or
individual validator object. This gives us the flexibility
of composing multiple validators without modifying this
class.